LayoutTests/storage/indexeddb/modern/workers-disabled-expected.txt

[Worker] Check to make sure IndexedDB in workers can be disabled at runtime

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/workers-disabled.js
PASS [Worker] self.indexedDB is undefined.
PASS [Worker] 'indexedDB' in self is false
PASS [Worker] self.IDBCursor is undefined.
PASS [Worker] 'IDBCursor' in self is false
PASS [Worker] self.IDBCursorWithValue is undefined.
PASS [Worker] 'IDBCursorWithValue' in self is false
PASS [Worker] self.IDBDatabase is undefined.
PASS [Worker] 'IDBDatabase' in self is false
PASS [Worker] self.IDBFactory is undefined.
PASS [Worker] 'IDBFactory' in self is false
PASS [Worker] self.IDBIndex is undefined.
PASS [Worker] 'IDBIndex' in self is false
PASS [Worker] self.IDBKeyRange is undefined.
PASS [Worker] 'IDBKeyRange' in self is false
PASS [Worker] self.IDBObjectStore is undefined.
PASS [Worker] 'IDBObjectStore' in self is false
PASS [Worker] self.IDBOpenDBRequest is undefined.
PASS [Worker] 'IDBOpenDBRequest' in self is false
PASS [Worker] self.IDBRequest is undefined.
PASS [Worker] 'IDBRequest' in self is false
PASS [Worker] self.IDBTransaction is undefined.
PASS [Worker] 'IDBTransaction' in self is false
PASS [Worker] self.IDBVersionChangeEvent is undefined.
PASS [Worker] 'IDBVersionChangeEvent' in self is false
PASS successfullyParsed is true

TEST COMPLETE
